fix(mu-helper): prevent ground item desync during helper/manual pickup

+++ src/source/MUHelper/MuHelper.cpp
@@ -1047,8 +1047,12 @@ namespace MUHelper
             }
             else
             {
-                SocketClient->ToGameServer()->SendPickupItemRequest(m_iCurrentItem);
-                DeleteItem(m_iCurrentItem);
+                if (SendGetItem == -1)
+                {
+                    SendGetItem = m_iCurrentItem;
+                    SocketClient->ToGameServer()->SendPickupItemRequest(m_iCurrentItem);
+                    DeleteItem(m_iCurrentItem);
+                }
             }
         }
